requirements.txt
numpy>=1.21.0
sounddevice>=0.4.4
soundfile>=0.10.3
scipy>=1.7.0
pillow>=9.0.0  # For icon creation 